requirements.txt
deepgram-sdk
google-generativeai
streamlit
sounddevice
numpy
python-dotenv
snowflake-connector-python
websockets
